+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 
 ### v2.2.19 / 未发布
 
+* [214](https://github.com/Vanessa219/vditor/issues/214) wysiwyg heading id `引入特性`
 * [206](https://github.com/Vanessa219/vditor/issues/206) combined blockquote and lists (windows firefox) `修复缺陷`
 * [151](https://github.com/Vanessa219/vditor/issues/151) 支持隐藏编辑器工具栏 `引入特性`
 * [121](https://github.com/Vanessa219/vditor/issues/121) 所见即所得模式支持 [ToC] `引入特性`